Add legacy Python code for feeding APE.
There are multiple sources for this code; it exists in a public personal
git repository and on our collaborative SURFdrive, at, respectively:

    https://github.com/simonscheider/QuAnGIS/tree/master/WorkflowSynthesis
    https://surfdrive.surf.nl/files/index.php/apps/files/?dir=/Shared/QuAnGIS/infrastructure/architecture/workflowspecifications/Preprocessing

At the moment of writing, they are mostly, though not exactly, the same,
modulo line endings (DOS vs Unix). I suggest that this repository will
hold the canonical version of these files from now on.