Dissolve Ontology/Taxonomy class distinction.
While it is useful to know when an ontology is also a Taxonomy, the
class distinctions we had now were only on the Python type level. They
were not validated and could be used more or less interchangeably. This
just complicated things, so I did away with it. We should just validate
the ontology as taxonomy when necessary.