Directly project types into SemTypes.
Projections are now computed directly into SemTypes, instead of
pre-generating the projection and then carrying it along to the chasms
where it happens to be most opportune to apply them. In my opinion, this
makes it a lot clearer what actually happens.

This also avoids us using SemTypes to carry along information that
happens to be structurally similar to a SemType but isn't actually a
SemType; that is, we used to compute each node's projection into a
"SemType" that only acted as a carrier for what the projection of that
node in each dimension.