Remove some code due to wrong assumptions.
Overloading is now not possible --- that is, it's still possible to
constrain variables in type signatures, but two wholly different type
signatures with the same name are no longer possible. I could
conceivably make this work again by making types immutable and
nondeterministically check every possible application. I'm not sure yet
if that is the correct way to go.

What lead to this first change is the second change: constraints will
now never lead to unification; they are only there as a check. Assuming
that they should lead to unification was a mistake on my part. In short,
given that we also have subtypes, unifying leads to issues.