Update tests so that they mostly work again.
Mostly. It's become increasingly clear to me that the approach of
explicitly indicating which functions can take a subtype is not
sufficient, because making everything a constrained variable whenever we
encounter issues only solves issues on an ad-hoc basis, possibly
introducing further issues. The following no longer works:

quangis.error.TypeMismatch: Error while applying:
	((fcont avg) ((interpol (pointmeasures temperature)) (deify (merge (pi2 (((select eq) (objectregions muni)) (object Utrecht))))))) (region x) : Itv such that x330 must be a subtype of Val, x330 must be a subtype of Val
	to
	ratio : (Ratio ** (Ratio ** Ratio))
Type mismatch. Could not unify:
	Itv with Ratio

This is because `pointmeasures` produces a `R2(Reg, Itv)`, and that Itv
is passed to `interpol ...`, resulting eventually in a `R2(Loc, Itv)`,
which is then passed to `fcont avg ...` to result in an `Itv`. But
`ratio` wants a `Ratio`. `Ratio` is a subtype of `Itv`, but not vice
versa. Before, `fcont` produced a `Ratio`, but now what it produces
depends on the input it receives. This is subtle: any function that
produces a `Ratio` of course also produces an `Itv`, because any `Ratio`
*is* an `Itv`. But the system doesn't know that unless we tell it, at
*any* point such an issue *could* arise.

I'm faily certain that the system can be improved in such a way that it
can transparently handle subtypes, and then our definitions can finally
be stable. It'll look more elegant, too. However, it would take some
time.