Change `tools:` prefix everywhere.
In commit 0898a9fed8a72604b67e5e758c9cca7cd9483619, I changed back a
prefix that prevented the code from working --- in confusion, since it
should not have worked before. I had completely missed that Enkhbold had
changed this prefix in 9c1cc8a67aa14640f62fa4311b7a7d1048a31407.

Whatever the reason, I have changed it back to the prefix introduced in
that commit, and have added changes to make the rest of the code not
fail.

It doesn't really matter: either URI is a work-in-progress anyway. The
proper, final URI should be more resistant to bit rot --- ie not be a
GitHub resource OR a webpage without a version number.