Add vocabulary to tool ontology.
I found the dichotomy between 'concrete' and 'abstract' tools quite
confusing, because it suggested that the union of these two captures
everything, which it doesn't: a supertool was neither abstract nor
concrete.

Furthermore, in our tool ontology, which class a tool falls under was
only implied, which made testing quite difficult (how do you test that
everything of type X satisfies conditions Y, if conditions Y are exactly
what tells you that something is of type X in the first place?)

That is why I have added a bit of vocabulary to the tool ontology and
indicated for every tool whether it is an implementation, a
specification or a combination (a supertool). I'm not sure whether these
are the most intuitive terms, but we do need *some* terms.