Remove packaging.
I don't want to deal with maintaining a general-purpose Python interface
to APE, because to do it properly, I'd have to address all APE features
with Pythonic structures, and I'd have to output Workflow structures
that aren't so particular to our project. I'll merge the single `ape.py`
file back into the <https://github.com/quangis/workflow-synthesis>
repository. It should be enough for anyone looking into how to interface
with APE to get a headstart.