Update CCT of LoadCountAmounts.
In the meeting, we figured out that a workflow was generated that feeds
a source of type  R(Obj, Reg * Nom) into the second input of a
ZonalStatisticsMeanInterval, the output which in turn gets passed into a
LoadCountAmounts along with a second use of the same source, now
interpreted as a R(Obj, Count). This is another instance of the issue
described in https://github.com/quangis/transforge/issues/79: we really
want to sometimes describe multiple attributes of the same thing.

See also
https://github.com/quangis/transforge/issues/79#issuecomment-1210661153

We now work around it by being consistent with the annotation.