tools.json now uses root of dimensions.
That is, the tools (as given to APE) provide their CCD type in all three
dimensions: if a dimension is empty, it is set to its root. However, the
APE *run* configuration, as extracted from questions, may just drop a
dimension if it is empty.

This makes sense: if we're looking for a workflow, we want to be as
specific as possible for our types, and if we can't be, then anything
goes.

But the workflow itself should only combine specific types, otherwise
we'd get in the situation that, when we ask for, say, a Ratio, we can
apply any tool that doesn't specify its type along the Nominal
dimension, and indeed, that was what happened.

The effect is that APE no longer finds workflows for a number of
questions.