`isOutputData=false` seems to reproduce CLI run.
We're having an APE issue that I just solved (I think) without really
understanding what's going on. We're on APE 1.1.12 and I noticed that
the command-line interface succeeds in giving us some solution workflows
where the Java API doesn't.

I figured out that I can make the Java API also provide solutions if I
change what I pass to `APERunConfig.withProgramInputs()` and
`.withProgramOutputs()`.

That is, we have types in JSON format (ie `{"Dimension":["Class",...]}`)
and turn them into `nl.uu.cs.ape.sat.models.Type` objects before giving
them as arguments to the aforementioned methods. To do this, we use
`Type.taxonomyInstanceFromJson`, which I assume is correct.

This method takes a boolean parameter `isOutputData`. I naively set it
to `true` for the outputs and `false` for the inputs, but this seems to
be causing the discrepancy between the CLI and the API; if I set them
both to `false`, we get solutions again.

What does the boolean do? Is always setting it to `false` reasonable?
(It occurs to me now that the boolean may be meant to differentiate
between "uninstantiated" types of tool descriptions, and type instances,
i.e. the "applied" types you're looking for in a workflow. Is that what
it is?)